reats/requirements.txt
boto3 == 1.28.49
cryptography == 41.0.7
django==4.1
django-extensions==3.2.3
django-phonenumber-field[phonenumberslite]==7.1.0
djangorestframework-simplejwt[crypto]==5.3.1
django-stubs[compatible-mypy]==4.2.3
djangorestframework==3.14.0
djangorestframework-stubs[compatible-mypy]==3.14.2
freezegun==1.2.2 
gunicorn==20.1.0
pillow==10.0.1
pre-commit==3.3.3
psycopg2==2.9.6
pytest-django==4.5.2
python-dotenv==1.0.0
